backend/pyproject.toml
[project]
name = "movie-recommender-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "pandas>=2.2",
    "numpy>=1.26,<2",
    "scikit-surprise>=1.1.3",
    "scikit-learn>=1.4",
    "scipy>=1.13",
    "psycopg2-binary>=2.9",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
]

[dependency-groups]
dev = ["pytest>=8", "httpx>=0.27"]

[tool.setuptools.packages.find]
exclude = ["data*"]
